- Duke: Move RRRA vehicle tilting out of the input code and into the playsim.

+++ source/games/duke/src/types.h
@@ -299,7 +299,7 @@ struct player_struct
 	int SeaSick;
 	short MamaEnd; // raat609
 	short moto_drink;
-	float TiltStatus, oTiltStatus;
+	short TiltStatus, oTiltStatus;
 	double VBumpNow, VBumpTarget;
 	short TurbCount;
 	short drug_stat[3]; // raat5f1..5
